Intermediate commit MOD Change Analysis to use the calibration     All results are saved in a new Result object which contains the     x and y position and the intensity value MOD MAJOR     The IdentMap and ByteMap pass to the Analysis method must NOW have     the same size. This prevents the need to calculate the rho and theta     position over and over again
    To prevent issues in Exp, the source pattern map and the source Hough
    map are changed if the dimensions of the map returns by the operation
    are different than from the original. In other words, the source map
    are changed under the binning and crop operations.

+++ src/org/ebsdimage/core/exp/ops/detection/results/Area.java
@@ -17,11 +17,6 @@
  */
 package org.ebsdimage.core.exp.ops.detection.results;
 
-import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.average;
-import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.max;
-import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.min;
-import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.standardDeviation;
-
 import org.ebsdimage.core.exp.Exp;
 import org.ebsdimage.core.exp.OpResult;
 
@@ -30,6 +25,10 @@ import rmlimage.core.BinMap;
 import rmlimage.core.IdentMap;
 import rmlimage.core.Identification;
 import rmlimage.module.real.core.RealMap;
+import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.average;
+import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.max;
+import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.min;
+import static org.ebsdimage.core.exp.ops.detection.results.ResultsHelper.standardDeviation;
 
 /**
  * Result operation that evaluate the area over peaks.
@@ -49,7 +48,7 @@ public class Area extends DetectionResultsOps {
 
         // ========= Calculate area ===========
 
-        float[] areas = Analysis.getArea(identMap).val;
+        double[] areas = Analysis.getArea(identMap).val;
 
         // ========= Calculate results ===========
